requirements.txt
asgiref==3.5.2
black==22.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.5
django-filter==21.1
django-oauth-toolkit==2.0.0
djangorestframework==3.13.1
drf-social-oauth2==1.2.1
ecdsa==0.17.0
filelock==3.7.1
idna==3.3
importlib-metadata==4.11.4
jwcrypto==1.3.1
Markdown==3.3.7
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
pipenv==2022.6.7
platformdirs==2.5.2
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.4.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
wrapt==1.14.1
zipp==3.8.0
